Split Collaboration Protocol into multiple files
